Removing django-crispy-forms and django-floppyforms
Resolve #433 #343
