+++ djadmin2/themes/djadmin2theme_bootstrap3/templates/djadmin2theme_bootstrap3/auth/password_change_form.html
@@ -1,5 +1,5 @@
 {% extends "djadmin2theme_bootstrap3/base.html" %}
-{% load i18n admin2_tags crispy_forms_tags %}
+{% load i18n admin2_tags %}
 
 {% block page_title %}{% trans "Password change" %}: {{ form.user }}{% endblock page_title %}
 
@@ -24,7 +24,7 @@
 
             <form method="post">
                 {% csrf_token %}
-                {{ form|crispy }}
+                {{ form }}
                 <button class="btn btn-sm btn-success" type="submit" name="_save">{% trans "Change my password" %}</button>
             </form>
         </div>
