+++ djadmin2/themes/djadmin2theme_bootstrap3/templates/djadmin2theme_bootstrap3/edit_inlines/stacked.html
@@ -1,8 +1,8 @@
-{% load i18n admin2_tags crispy_forms_tags %}
+{% load i18n admin2_tags %}
 
 {% for inline_form in formset %}
     <div class="change_form space-below">
-        {{ inline_form|crispy }}
+        {{ inline_form }}
         {% if not inline_form.visible_fields %}
             <p class="alert alert-warning empty-form">
                 {% trans "This form doesn't have visible fields. This doesn't mean there are no hidden fields." %}
