+++ docs/installation.rst
@@ -23,8 +23,6 @@ Add djadmin2 and rest_framework to your settings file:
         'djadmin2',
         'djadmin2.themes.djadmin2theme_bootstrap3', # for the default theme
         'rest_framework', # for the browsable API templates
-        'floppyforms', # For HTML5 form fields
-        'crispy_forms', # Required for the default theme's layout
         ...
     )
 
@@ -32,7 +30,6 @@ Add djadmin2 and rest_framework to your settings file:
         'DEFAULT_PAGINATION_CLASS': 'rest_framework.pagination.PageNumberPagination',
         'PAGE_SIZE': 10
     }
-    CRISPY_TEMPLATE_PACK = "bootstrap3"
     ADMIN2_THEME_DIRECTORY = "djadmin2theme_bootstrap3"
 
 Add djadmin2 urls to your URLconf:
@@ -70,5 +67,4 @@ Themes are a new default theme based on bootstrap3 and also some new settings to
         'DEFAULT_PAGINATION_CLASS': 'rest_framework.pagination.PageNumberPagination',
         'PAGE_SIZE': 10
     }
-    CRISPY_TEMPLATE_PACK = "bootstrap3"
     ADMIN2_THEME_DIRECTORY = "djadmin2theme_bootstrap3"
