+++ setup.py
@@ -131,9 +131,7 @@ setup(
         'django-extra-views>=0.6.5',
         'django-braces>=1.3.0',
         'djangorestframework>=3.3.3',
-        'django-floppyforms>=1.6.2',
         'django-filter>=0.13.0',
-        'django-crispy-forms>=1.3.2',
         'pytz==2014.7',
         'future>=0.15.2',
         ],
