+++ tox.ini
@@ -6,11 +6,11 @@ exclude = migrations/*,docs/*
 
 [tox]
 envlist =
-    py27-{1.8,1.9,master},
-    py33-{1.8},
-    py34-{1.8,1.9,master},
-    py35-{1.8,1.9,master},
-
+;    py27-{1.8,1.9,master},
+;    py33-{1.8},
+;    py34-{1.8,1.9,master},
+;    py35-{1.8,1.9,master},
+    py35-1.9
 [testenv]
 commands =
     flake8 djadmin2
